changes in CypressFX2.java and CypressFX2MonitorSequencer.java to support new timestamp tick, synchronisation and timestamp resetting with reset events
CypressFX2.java branches on DID whether to use new scheme or not.

timestamp tick and master/slave state are now displayed in titlebar

dollbrain: added "pixels" for face features and face detector

git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@405 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/caviar/hardwareinterface/usb/CypressFX2Mapper.java
@@ -32,9 +32,9 @@ public class CypressFX2Mapper extends CypressFX2MonitorSequencer{
      * prints trigger mode the screen
      * @return timestamp tick in microseconds
      */
-    public double getOperationMode() throws HardwareInterfaceException
+    public float getOperationMode() throws HardwareInterfaceException
     {
-        double tick;
+        float tick;
         
          if (!isOpen())
         {
@@ -73,19 +73,19 @@ public class CypressFX2Mapper extends CypressFX2MonitorSequencer{
         if (dataBuffer.Buffer()[1]==0x00)
         {
             log.info("Trigger mode: Host (Master)");
-            tick=1;
+            tick=1f;
         } else if (dataBuffer.Buffer()[1]==0x01)
         {
             log.info("Trigger mode: Host (Master)");
-            tick=0.125/6;
+            tick=(float)0.125/6;
         } else if (dataBuffer.Buffer()[1]==0x02)
         {
             log.info("Trigger mode: Slave");
-            tick=1;
+            tick=1f;
         } else if (dataBuffer.Buffer()[1]==0x03)
         {
             log.info("Trigger mode: Slave");
-            tick=0.125/6;
+            tick=(float)0.125/6;
         } else
         { 
             log.warning("invalid tick: "+ dataBuffer.Buffer()[1]);
